[RED]: Roll.countBySide fails to count 2 Side

+++ java/src/main/java/org/yatzykata/valueobject/Roll.java
@@ -32,7 +32,7 @@ public class Roll {
             '}';
     }
 
-    public int countBySide(Side side) {
+    public long countBySide(Side side) {
         return 1;
     }
 }
